Resolve issue 14
`Status` and `Name` were improperly being read due to new types of
values in the text product headers. In this case,
"POST-TROPICAL CYCLONE" and "REMNANTS OF" which previously were not
recognized. These values have been added to list of options as
acceptable. Additionally, "REMANTS OF" should be cleaned and produce
only "REMNANTS" as `Status` now in returned dataframe.